elaborate: empty blocks always get a 1-byte pad pseudo-register
Structs with zero members are:
 - zero length in GCC even though that’s not strictly allowed in the C spec
    - this is a GCC-only extension which allows this and makes them zero length
 - length 1 in G++ since C++ strictly forbids zero size objects

This unexpected discrepancy in sizes of the structs results in blocks
in the regmap slowly slipping by 1 byte for each earlier zero-member
struct when compiled as C++.  Structs end up misaligned in the memory
map.

Always insert a 1-byte pad member in every struct which would
otherwise have zero members.