regmap: Python library for programmatically specifying a register map
This library provides a set of base classes for use in capturing a hierarchical
register structure and mapping it onto an address space. A register hierarchy is
defined by means of sub-classing generalized base classes for various object
types and attaching these sub-classes as attributes within a parent class. All
configuration data are passed via keyword arguments to the sub-class definition.

Upon instantiation, a specification will recursively instantiate all of it's
sub-class members and link them into an internal tree structure. This tree will
form the basis from which programmatic IO access to registers will be provided
(refer to subsequent commits in this series).