regmap: Extend regio-generate to emit a regmap specification
This enhancement allows regio-generate to translate an elaborated YAML IR input
file to a regmap specification by emitting Python files containing sub-classes
for each block, register and field. The blocks are pulled into an address space
to define the top-level map for each PCIe BAR found in the IR. The module for
the top-level can be imported to instantiate the regmap and IO for the BARs,
wrap them in a proxy and make use of them for accessing registers.