regmap: Create a regio library to allow exporting tools along with regmap
This change makes regmap a sub-package (in the Python sense) of a new top-level
regio package. This allows distributing the existing regio* tools in a way that
is consistent with standard Python distribution pactices. In-tree regio* stubs
have been created to provide backwards compatibility (the stubs are not used by
the distribution). All jinja2 templates are included in the distribution and are
installed along side the Python sources.

The source tree has been shuffled to conform to the distribution tool's needs.