regmap: Add support specifying object paths to the dump command
This change extends the dump command with the ability to provide object paths
to display. This is a convenient alternative to using eval, since it will
automatically create a variable on the given paths. Note that unlike eval, the
paths are not interpreted as Python code, but rather split up on the '.' and
looked up.

The following:
regio-esnet-smartnic dump dev0.bar2

is equivalent to:
regio-esnet-smartnic eval 'dev0.bar2()'