regmap: Support for proxy groups
This change introduces the ability to group multiple regmap nodes of the same
type and treat them as one unit. The current use case is for slicing an array,
where each element selected by the indexing key can be operated on as a whole
rather than writting looping code to perform the same actions.