regmap: Add support for displaying object paths as relative or absolute
This change extends the variable formatting output to allow paths to be
displayed relative to the object the variable was created on. All nodes
under the hierarchy of the variable will have a path relative to the root.
The default as been changed to relative paths, with an option to enable
absolute paths via either the command line --abspath flag or by passing the
abspath=True keyword when invoking the proxy to create variable.

- Via command line option:
regio-esnet-smartnic --abspath eval 'dev0.bar2()'

- Via keyword argument:
regio-esnet-smartnic eval 'dev0.bar2(abspath=True)'