regmap: Fix generated Python blocks to properly handle register width
Registers without fields declared with a width of 8 or 64 bits were not being
properly emitted. This change updates the generator such that registers with
less than the data width will be given an implicit field, where as registers
of an exact multiple of the data width will be sized correctly.