regmap: Don't attempt reading write-only registers/fields during display
This change respects access permissions when displaying registers and fields.
Note that this currently only applies when displaying. Access permissions are
not enforced on reads/writes performed by proxy objects in order to allow for
building tests of error paths and edge cases.

Also included a new column in verbose mode to show the access.