elaborate: Introduce new "enum" attribute for fields to replace "enum_hex"
The "enum_hex" attribute currently allows both strings and integers for
specifying the value of an enumeration. This results in the following
scenario:

--- YAML start ---
enum_hex:
  a:    first  # meant to be 0xa, but parsed as the string "a"
  10:   second # meant to be 0x10, but parsed as the base-10 integer 10
  0x10: third  # meant to be 0x10, but parsed as the base-16 integer 16
--- YAML end ---

Due to the way the enumeration value is used by the source generators, the
code generated for "a" and 10 will be emitted correctly with values 0xa and
0x10 (because the integer 10 will be converted to a string). However, 0x10
will be emitted incorrectly with value 0x16 (also due to string conversion).

The new "enum" attribute will enforce that the enumeration values used as the
mapping keys always be integers.

The "enum_hex" attribute is now deprecated and will be removed in the future.
Uses of "enum_hex" will produce warnings during elaboration.